+++ src/yt_session.c
@@ -41,7 +41,7 @@ session_planet_offset(const struct yt_session *session)
 }
 
 void
-session_set_pager_line_count(struct yt_session *session, float value)
+session_set_pager_line_count(struct yt_session *session, int value)
 {
 	session->pager.line_count = value;
 }
