+++ tests/test_data.c
@@ -1449,7 +1449,7 @@ test_file_viewer_physical_stream(void)
 	struct yt_error error;
 	int foreground = 5;
 	bool bold = false;
-	float line_count = 19.0f;
+	int line_count = 19;
 	int pager_foreground = 5;
 	char key[2] = "";
 
@@ -1485,7 +1485,7 @@ test_file_viewer_physical_stream(void)
 	    && viewer.lengths[2] == 8U
 	    && memcmp(viewer.rows[2], "  - item", 8U) == 0
 	    && foreground == 5 && pager_foreground == 5
-	    && bold && line_count == 0.0f);
+	    && bold && line_count == 0);
 	CHECK(yt_file_delete(actual, false, &error));
 #ifdef _WIN32
 	_rmdir(directory);
